skills/storage/pds/alibabacloud-pds-intelligent-workspace/scripts/requirements.txt
requests==2.32.2
python-pptx==0.6.23
Pillow==12.1.1